♻️ refactor: BookBtn 컴포넌트 상위로 이동
